requirements.txt
python-telegram-bot==13.7
requests==2.26.0
pyrogram==2.0.106
twilio==8.5.0